Add cast operations in load/store.
Due to imprecisions of the heap abstraction the bitwidth of a region
and the bitwidth of the value to be read from or write to that region
might be different. For now, we add sext/trunc operations to make sure
the resulting Crab CFG is well typed. However, we might need to ignore
the load/store in this case since I am not sure what it means if the
truncate operation (needed during the store) overflows.