[NEW] resolve indirect calls using LLVM-Dsa.
Indirect calls can be resolved using the option --devirt-functions=[none|types|dsa]
- none : do not resolve indirect calls (default)
- types: select all functions with same type signature
- dsa  : use Dsa analysis to select the callees

The option --devirt-functions=dsa is potentially the most precise
since it uses the LLVM Dsa pointer analysis to figure out the possible
callees of an indirect call. However, sometimes when either running
the pointer analysis is expensive or precision is not really important
the option --devirt-functions=types can be used instead. This option
resolves the indirect call by simply considering all functions which
match their type signature with the callsite.