Fixes in the heap translation:
1) preserve order of regions between callsites and functions declarations.

2) resolve disagreements between callsite caller and callee.  It is
possible at a particular callsite that a callee cell is mapped to a
caller cell such that they do not agree on all details that the heap
translation cares about. This is something should probably be fixed in
sea-dsa. Meanwhile, we make sure that a function is consistent with
all its callsites in terms of memory regions.