Generate same unconstrained Crab array index.
Note that we completely ignore the index of Crab array reads and
writes.  This is OK for array smashing but it is not acceptable for
other array domains that we plan to integrate in the near future.

We plan to perform some LLVM static analysis so that we can at
translation time to infer the integer offset of a pointer wrt to its
allocation site.