Flag to disable lowering of vector operations into scalars.
Clam does not understand vector operations so we try to avoid them.

By default, clam.py tries to make sure that Clang does not use vector
operations using flags fno-vectorize and fno-slp-vectorize.  However,
this is not enough if the bitcode is generated by external Clam
clients. For this, Clam uses a new LLVM transformation to lower
vector operations into scalars.