Use array_store_range instead of array_init if no smashing
These two Crab statements are very similar but their semantics differ
in a very subtle way. The former is just a sequence of Crab
array_store instructions. The latter is similar to SMT constant array
which means that a call to array_init with array A always kill any
existing cells of A.

If we have an array of struct of integers we will not have a single
call to array_init for the whole array but instead, we might have
multiple calls to array_init. With array adaptive domain multiple
calls to array_init will kill previous cells which is imprecise.

Note that for array_smashing domain we need array_init semantics. A
sequence of array_store will not be precise due to weak updates.