refactor: support for new crab API.
This the first of many more commits. All the code related to
inter-procedural and path analysis is commented. The code for the
intra-procedural analysis compiles and it is runnable although no tests
pass at this point.

The bottom-up inter-procedural analysis and the code related to
translate preserving memory SSA form have been removed because we do
not want to maintain them.