refactor: remove support for ShadowMem
We would like to have the option of extracting all memory region
information from sea-dsa ShadowMem, specially now that it has an API
to be used directly from LLVM MemorySSA. However, the support for
ShadowMem has been always partial (only intra-procedural) and the code
really needs to be rewritten.

The initial support was motivated by the need of the SeaHorn
Path-based BMC engine to use the translation from LLVM to Crab
preserving memory SSA form. However, it seems that we don't need that
anymore.

Thus, I prefer not to maintain this code unless is needed again.