fix(linker): gnu ld is more restrictive than clang linker.
I tried to isolate the code that instantiates each abstract domain and
move it to a different cpp file. This works using the clang ldd linker
but not with gnu ld. I do not want to fight with the linker right now
so the solution is that all the domain instantiations occur in Clam.cc
via macro REGISTER_DOMAIN.