refactor(cfg-builder): consider globals when purify a function
During function purification, we were considering the contents of
globals but not the address of a global.  With this commit, we also
pass as __input__ argument all the globals that are accessed by a
function or its callees.

We ask sea-dsa to know which variables are accessed by a function and
its callee.