refactor(cfgBuilder): CFG is built in two phases.
First, we create the a Crab function declaration for each LLVM
function and create an entry and an exit block. Second, we build the
rest of the CFG. We want a two-step approach because then at a
callsite we want to check for type mismatch errors between callee and caller
and more importantly, we might need to add cast statements between
regions in cases a caller passes an unknown (e.g., collapsed) region
but at the callee is not unknown.