fix(builder): insert correctly region_init statements.
I hopefully fix this once and for all. We use seadsa to figure out
which regions should be initialized by a function. Intuitively, a
function f should initialize region rgn if rgn is originated from some
allocated memory in f.  Note that even if f does not use rgn we still
want f to initialize rgn.

For this, two conditions must be satisfied: (1) region is unreachable
from input parameters or globals (except main), and (2) cannot be
returned by a callee unless the region was already an input to the
callee (i.e, cannot be an output parameter of a callsite).