fix(sea-dsa-wrapper): create a region if modified/read but not types.
In seadsa, we keep track of the types of the data for each pointer
dereference. If not types available then we were assuming that the
pointer is never dereferenced and therefore, we don't generate a
region for it. However, seadsa allows users to have special
annotations such as sea_dsa_set_modified. This will mark a seadsa node
as modified but without adding any type. This fix pretends that if the
marker is modified or read but no actual dereference happened (i.e.,
node types are empty) then the node was dereferenced at offset 0.