refactor(cfgBuilder): removed use of shared_ptr.
This commit changes the API of clam a bit. The typedef CfgBuilderPtr
is not available anymore.

  using CfgBuilderPtr = std::shared_ptr<clam::CfgBuilder>;

The manager (CfgBuilderManager) owns CfgBuilder so it keeps it as a
unique_ptr.  When asked the manager about a CfgBuilder, a reference or
raw pointer is returned so ownership is not affected in any way.