feat: instrumenting properties at CrabIR level
New class CrabIREmitter that describes the API that must be
implemented in order to instrument a property. The API consists of
visit_X methods where X is (for now) a memory-related instruction. The
CFGBuilder calls to each of these visit methods before and after the
corresponding CrabIR statement has been added.

The goal of this API is to allow Clam clients to add instrumentation
without modifying Clam code.

As examples, we have implemented an instrumentation that adds
null-dereference checks (lib/Clam/Properties/NullCheck) and another
one that adds use-after-free checks
(lib/Clam/Properties/UafCheck).

The option --crab-check from clam.py takes (among others) as possible
values null, null-legacy, uaf, and uaf-legacy. The legacy
instrumentations are done at the LLVM IR level while the non-legacy
are the new ones done at the CrabIR level. The legacy variants will be
removed but they are kept for now for debugging purposes.