feat(cfg): replace unsigned comparison with signed comparisons
Clam has already an option that replaces an unsigned comparison with
several signed ones. This replacement is performed by a LLVM
transformation. However, sometimes, it is desirable not to modify the
LLVM bitcode and instead, perform the replacement at the CrabIR
level. This commit adds a new option in CfgBuilderParams to remove
unsigned comparisons directly at the CrabIR level without modifying
the LLVM bitcode.