feat(cfg-builder): support for arithmetic with overflow intrinsics
If flag lower_arithmetic_with_overflow_intrinsics (defined in
CrabBuilderParams) is enabled then the arithmetic intrinsic is
replaced with a standard arithmetic operation and it assumes that the
operation does not overflow. The translations takes place when an
ExtractValueInst is used to access to the pair returned by the
arithmetic with overflow intrinsics. If the ExtractValueInst indexes
the first element of the pair then the result of the arithmetic
operation is returned. If the ExtractValueInst indexes the second
element of the pair then the result is always false (i.e., no
overflow).