refactor(bounds): unneccesary assign to propagate size/offsets
It is important to minimize the number of assignments in the abstract
domain. Currently, each gep adds two assignments to progagate sizes
and offsets, respectively. However, many gep have zero offsets so we
can reuse the offset variable from the pointer operand. Moreover, we
do not need to add assignments to progagate sizes since we can also
reuse the size variable from the pointer operand.