refactor(optimizer): mark function as InaccessibleMemOnly
LLVM may remove all calls to verifier.assume if it is marked as
ReadNone or ReadOnly even if the function is marked as OptimizeNone.
The attribute InaccessibleMemOnly should be good enough because it
ensures that the call may only access to memory that is innaccesible
from the IR. Thus, LLVM optimizations should be okay.