feat(domains): new Octagons domain soct
This domain was available in Crab but not in Clam.  The domain is as
expressive as Elina/Apron Octagons but it keeps the DBMs in a special
form called Split Normal Form which can have sometimes better
performance. Implementation details are described in the paper A Fresh
Look at Zones and Octagons
(https://jorgenavas.github.io/papers/ACM_TOPLAS_zones_octagons.pdf)