refactor(clam-diff): more improvements and fixes
- chore(clam-diff): removed debug code
- fix: dummy json read/write if boost < 1.80
- fix(json): parsing of true/false
- perf(clam-diff): use of caching to avoid parsing same invariants

The use of caching helps us in two ways. First, we do not convert the
same string to linear_constraint_system_t over and over
again. Moreover, since the cache is used by the two calls to
read_json, if the two domains produce the same invariant string then
we do not need to call the abstract domain for comparison.

- fix(clam-diff): (if enabled) exclude ghost variables semantically

missed in previous commit